+++ test/jellyfish/sdk_test.exs
@@ -1,20 +1,22 @@
 defmodule Jellyfish.SDKTest do
   use ExUnit.Case
 
-  alias Jellyfish.SDK
+  alias Jellyfish.SDK.Client
 
   @url "https://somemockurl.com"
 
   describe "sdk" do
     test "creates client struct" do
-      client = SDK.new(@url)
+      client = Client.new(@url)
 
-      assert %Tesla.Client{
-               adapter: {Tesla.Adapter.Hackney, :call, [[]]},
-               pre: [
-                 {Tesla.Middleware.BaseUrl, :call, [@url]},
-                 {Tesla.Middleware.JSON, :call, [[]]}
-               ]
+      assert %Client{
+               http_client: %Tesla.Client{
+                 adapter: {Tesla.Adapter.Hackney, :call, [[]]},
+                 pre: [
+                   {Tesla.Middleware.BaseUrl, :call, [@url]},
+                   {Tesla.Middleware.JSON, :call, [[]]}
+                 ]
+               }
              } = client
     end
   end
